+++ lib/async_or_error.mli
@@ -38,6 +38,9 @@ val of_exn : exn -> _ t
 val of_exn_result : ('a, exn) Result.t Deferred.t -> 'a t
 val error : string -> 'a -> ('a -> Sexp.t) -> _ t
 val error_string : string -> _ t
+val errorf : ('a, unit, string, _ t) format4 -> 'a
+val tag : 'a t -> string -> 'a t
+val tag_arg : 'a t -> string -> 'b -> ('b -> Sexp.t) -> 'a t
 val unimplemented : string -> _ t
 val combine_errors : 'a t list -> 'a list t
 val combine_errors_unit : unit t list -> unit t